requirements.txt
pytest
aiohttp
aioprocessing
nest-asyncio
numpy
pandas
tqdm
websockets
loguru
python-dotenv
eth-abi
web3
multicall
slither-analyzer
janus
# Django admin
asgiref
Django
django-jazzmin
psycopg2-binary
sqlparse
django-cors-headers
djangorestframework
Markdown
django-filter
redis
coverage